Refactor header to share base component
